// src/data/generated/rules.ts
export const FASTING: readonly FastingPeriod[] = [
	{
		cmd: "nday >= -48 && nday <= -9",
		rules: [
			{ level: "0000111", cmd: "dow == 0 || dow == 6" },
			{ level: "0000001", cmd: "dow != 6 && dow != 0" },
			{ level: "0000000", cmd: "nday == -48 || nday == -47" },
			{ level: "0000001", cmd: "nday == -46" },
			{ level: "0000000", cmd: "nday == -45" },
			{ level: "0000011", cmd: "nday == -44" },
			{ level: "0000001", cmd: "nday == -25" },
			{ level: "0000111", cmd: "nday == -18 || (nday == -20 && doy == 80)" },
			{ level: "0000111", cmd: "nday == -17 || (nday == -19 && doy == -81)" },
			{ level: "0000111", cmd: "doy == 82" },
			{ level: "0000111", cmd: "(dRank >= 4 || doy == 84) && nday >= -43" },
			{ level: "0000011", cmd: "(dRank >= 4 || doy == 84) && nday >= -43 && (dow == 3 || dow == 5)" },
			{ level: "0000010", cmd: "nday == -16" },
			{ level: "0011111", cmd: "dRank >= 6" },
		],
	},
	{
		cmd: "nday >= -8 && nday <= -1",
		rules: [
			{ level: "0001111", cmd: "nday == -8" },
			{ level: "0011111", cmd: "nday == -8 && dRank >= 6" },
			{ level: "0011111", cmd: "dow == 0" },
			{ level: "0000001", cmd: "(dow == 1 || dow == 2 || dow == 3)" },
			{ level: "0000111", cmd: "(dow == 1 || dow == 2 || dow == 3) && dRank >= 6" },
			{ level: "0000111", cmd: "dow == 4" },
			{ level: "0000000", cmd: "dow == 5" },
			{ level: "0000010", cmd: "dow == 5 && dRank >= 6" },
			{ level: "0000010", cmd: "nday == -1" },
			{ level: "0000111", cmd: "nday == -1 && dRank >= 6" },
		],
	},
	{
		cmd: "nday >= 0 && nday <= 56",
		rules: [
			{ level: "1111111", cmd: "!(dow == 3 || dow == 5)" },
			{ level: "0000111", cmd: "dow == 3 || dow == 5" },
			{ level: "0011111", cmd: "nday == 24 || nday == 38" },
			{ level: "0011111", cmd: "(dow == 3 || dow == 5) && dRank >= 4" },
			{ level: "1111111", cmd: "nday >= 0 && nday <= 7" },
			{ level: "1111111", cmd: "nday >= 49 && nday <= 56" },
		],
	},
	{
		cmd: "nday >= 57 && doy <= 179",
		rules: [
			{ level: "0000111", cmd: "dow == 2 || dow == 4" },
			{ level: "0000001", cmd: "dow == 1 || dow == 3 || dow == 5" },
			{ level: "0011111", cmd: "dow == 0 || dow == 6" },
			{ level: "0011111", cmd: "(dow == 1 || dow == 2 || dow == 4) && dRank >= 4" },
			{ level: "0000111", cmd: "(dow == 3 || dow == 5) && dRank == 4" },
			{ level: "0011111", cmd: "(dow == 3 || dow == 5) && dRank >= 5" },
			{ level: "0011111", cmd: "doy == 174" },
			{ level: "1111111", cmd: "doy == 179" },
			{ level: "0011111", cmd: "doy == 179 && (dow == 3 || dow == 5)" },
		],
	},
	{
		cmd: "doy >= 212 && doy <= 226",
		rules: [
			{ level: "0000001", cmd: "dow == 1 || dow == 3 || dow == 5" },
			{ level: "0000011", cmd: "dow == 2 || dow == 4" },
			{ level: "0000111", cmd: "dow == 6 || dow == 0" },
			{ level: "0000111", cmd: "dRank >= 4 && !(dow == 6 || dow == 0)" },
			{ level: "0011111", cmd: "doy == 217" },
			{ level: "1111111", cmd: "doy == 226" },
			{ level: "0011111", cmd: "doy == 226 && (dow == 3 || dow == 5)" },
		],
	},
	{
		cmd: "doy >= 318 && doy <= 353",
		rules: [
			{ level: "0000001", cmd: "dow == 1 || dow == 3 || dow == 5" },
			{ level: "0000111", cmd: "dow == 2 || dow == 4" },
			{
				level: "0000111",
				cmd: "(dow == 1 || dow == 3 || dow == 5) && (doy == 319 || doy == 328 || doy == 333 || doy == 337 || doy == 338 || doy == 339 || doy == 342 || doy == 350 || doy == 353)",
			},
			{
				level: "0011111",
				cmd: "!(dow == 1 || dow == 3 || dow == 5) && (doy == 319 || doy == 328 || doy == 333 || doy == 337 || doy == 338 || doy == 339 || doy == 342 || doy == 350 || doy == 353)",
			},
			{ level: "0011111", cmd: "doy == 324" },
			{ level: "0000001", cmd: "doy == 318" },
			{ level: "0011111", cmd: "dow == 6 || dow == 0" },
		],
	},
	{
		cmd: "doy >= 354 && doy <= 357",
		rules: [
			{ level: "0000001", cmd: "dow == 1 || dow == 3 || dow == 5" },
			{ level: "0000111", cmd: "dow == 2 || dow == 4" },
			{ level: "0000111", cmd: "dow == 6 || dow == 0" },
			{ level: "0000111", cmd: "doy == 357" },
		],
	},
	{
		cmd: "(doy >= 358 && doy <= 365) || doy <= 5",
		rules: [{ level: "1111111" }, { level: "0000111", cmd: "doy == 4" }],
	},
	{
		cmd: "(doy >= 6 && nday <= -56) || (doy >= 180 && doy <= 211) || (doy >= 227 && doy <= 317)",
		rules: [
			{ level: "1111111", cmd: "!(dow == 3 || dow == 5)" },
			{ level: "0000001", cmd: "dow == 3 || dow == 5" },
			{ level: "0011111", cmd: "(dow == 3 || dow == 5) && dRank >= 5" },
			{ level: "0000111", cmd: "(dow == 3 || dow == 5) && (dRank == 3 || dRank == 4)" },
			{ level: "0000011", cmd: "(dow == 3 || dow == 5) && dRank == 2" },
		],
	},
	{
		rules: [
			{ level: "1111111", cmd: "nday >= -69 && nday <= -63" },
			{ level: "0111111", cmd: "nday >= -55 && nday <= -49" },
			{ level: "0000111", cmd: "doy == 256" },
			{ level: "0000111", cmd: "doy == 240" },
		],
	},
];
// ...
export const DIVINE_LITURGY_COMMANDS: readonly LiturgyCommand[] = [
	{
		name: "Transfer",
		value: "nday >= 52 || nday < -55",
		cmd: "GS == 1",
		comment: "This determines whether the sequential readings may be skipped or transferred.",
	},
	{
		name: "Transfer",
		value: "nday >= 52 || nday < -55",
		cmd: "GS == 0",
		comment: "This determines whether the sequential readings may be skipped or transferred.",
	},
	{
		name: "Suppress",
		value: "doy == 4 && dow != 0 && dow != 6",
		cmd: "GS == 1",
		comment: "Eve of Theophany can have seq. readings on Sat. (see Table 25b, 2019) and on Sun. (see Table 24b, 2087)",
	},
	{
		name: "Suppress",
		value: "doy == 357 && dow != 6",
		cmd: "GS == 1",
		comment: "Eve of Nativity can have seq. readings on Sat. (see Table 27a, 2084)",
	},
	{
		name: "Suppress",
		value: "(doy == 358 || doy == 5)",
		comment: "Nativity and Theophany never have seq. readings",
	},
	{
		name: "Suppress",
		value: "(doy == 256 || doy == 217) && dow == 0",
		comment: "Suppress seq. readings if Exaltation or Transfiguration falls on Sunday",
	},
	{
		name: "Class3Transfers",
		value: "!(doy == 358 || doy == 357 || doy == 4 || doy == 5 || doy == 256 || doy == 217) && ((dRank >= 5 && dow != 0) || dRank == 8)",
		cmd: "GS == 1 || GS == 0",
	},
	{
		name: "Class3Transfers",
		value: "(doy == 3 && dow == 5) || (doy == 356 && dow == 5)",
		cmd: "GS == 1 || GS == 0",
		comment: "There is no Liturgy on Friday when the Royal Hours occur. The readings are moved to Thursday.",
	},
	{
		name: "Class3Transfers",
		value: "(doy == 256 || doy == 217) && dow != 0",
		comment: "Exaltation and Transfiguration readings are only transferred if it's not a Sunday",
	},
	{
		name: "TransferRulesB",
		value: "dow != 0 && !(nday >= -63 && nday <= 0) && doy != 338",
		cmd: "GS == 1",
		comment: "This determines the rules used to transfer the readings to the previous day. Note that St. Nicholas's day has a special set of rules",
	},
	{
		name: "TransferRulesF",
		value: "!(nday >= -63 && nday <= 0) && (dow == 2  && doy != 339) || ((doy == 340 && (dow != 0 && dow != 2)))",
		cmd: "GS == 1",
		comment: "This determines the rules used to transfer the readings to the next day (forwards). Note that St. Nicholas's day has a special set of rules.",
	},
	{
		name: "TransferRulesB",
		value: "dow != 0 && !(nday >= -63 && nday <= 0) && doy != 338",
		cmd: "GS == 0",
		comment: "This determines the rules used to transfer the readings to the previous day. Note that St. Nicholas's day has a special set of rules",
	},
	{
		name: "TransferRulesF",
		value: "!(nday >= -63 && nday <= 0) && (dow == 2  && doy != 339) || ((doy == 340 && (dow != 0 && dow != 2)))",
		cmd: "GS == 0",
		comment: "This determines the rules used to transfer the readings to the next day (forwards). Note that St. Nicholas's day has a special set of rules.",
	},
];

export const SCRIPTURE_TRANSFER_COMMANDS: readonly LiturgyCommand[] = [
	{
		name: "Suppress",
		value: "(doy == 358 || doy == 357 || doy == 4 || doy == 5)",
		comment: "Suppress sequential readings on eve and day of Nativity and Theophany",
	},
	{
		name: "Suppress",
		value: "(doy == 355 || (doy == 356 && dow != 1) || (doy == 357 && (dow >= 3 || dow == 0)) || (doy == 358 && (dow >= 4 || dow == 0)) || (doy == 359 && (dow >= 5 || dow == 0)) || (doy == 360 && (dow == 6 || dow == 0)) || (doy == 361 &&  dow == 0)) && (ndayF >= -125 && ndayF <= -105)",
		cmd: "GS == 1",
		comment: "This suppresses the readings that occur over Christmas in the case that there are 3 repeats and it is the 31st week of effective readings.",
	},
	{
		name: "Suppress",
		value: "(doy == 362 || (doy == 363 && dow != 1) || (doy == 364 && (dow >= 3 || dow == 0)) || (doy == 361 && dow != 0) || (doy == 360 && (dow < 6 && dow != 0)) || (doy == 359 && (dow < 5 && dow != 0)) || (doy == 358 && (dow < 4 && dow != 0)) || (doy == 357 && (dow < 3 && dow != 0)) || (doy == 356 && (dow == 1))) && (ndayF >= -118 && ndayF <= -81)",
		cmd: "GS == 1",
		comment: "This suppresses the readings that occur over Christmas in the case that there are 2 repeats and it is the 32nd week of effective readings.",
	},
	{
		name: "Suppress",
		value: "((doy == 0 && (dow >= 4 || dow == 0)) || (doy == 1 && (dow >= 5 || dow == 0)) || (doy == 2 && (dow == 6 || dow == 0)) || (doy == 3 &&  dow == 0)) &&  (nday >= -118 && nday <= -81)",
		cmd: "GS == 1",
		comment: "This suppresses the readings that occur over Christmas in the case that there are at least 2 repeats and it is the 32nd week of effective readings and it is January.",
	},
	{
		name: "Suppress",
		value: "(doy == 4 || (doy == 5 && dow != 1) || (doy == 6 && (dow >= 3 || dow == 0)) || (doy == 7 && (dow >= 4 || dow == 0)) || (doy == 8 && (dow >= 5 || dow == 0)) || (doy == 9 && (dow == 6 || dow == 0)) || (doy == 10 &&  dow == 0)|| (doy == 3 && dow != 0) || (doy == 2 && (dow < 6 && dow != 0)) || (doy == 1 && (dow < 5 && dow != 0)) || (doy == 0 && (dow < 4 && dow != 0))) && (nday <= -78 && nday >= -111)",
		cmd: "GS == 1",
		comment: "This suppresses the readings for the case that there is only 1 repeat and it is the 33rd week of effective readings",
	},
	{
		name: "Suppress",
		value: "((doy == 364 && (dow < 3 && dow != 0)) || (doy == 363 && (dow == 1))) && (ndayF <= -78 && ndayF >= -111)",
		cmd: "GS == 1",
		comment: "This suppresses the readings for the case that there is at least 1 repeat and it is the 33rd week of effective readings and it is still the old year.",
	},
	{ name: "AcceptTransfers", value: "dow != 0 && dRank < 5 && !(nday < 0 && nday > -63)" },
	{ name: "TransferRulesB", value: "dow != 0 && dow != 1 && dRank >= 5 && !(nday < 0 && nday > -63)" },
	{ name: "TransferRulesF", value: "dow == 1 && dRank >= 5 && !(nday < 0 && nday > -63)" },
];